+++ example/lib/main.dart
@@ -26,6 +26,9 @@ class ExampleArcaneApp extends StatelessWidget {
           HomeScreen(),
         ],
         theme: ArcaneTheme(
+            gutter: GutterTheme(
+              gutterCalc: (inWidth) => inWidth * 0.25,
+            ),
             themeMode: ThemeMode.system,
             scheme: ContrastedColorScheme.fromScheme(ColorSchemes.zinc)),
       );